refactor: Move chatroom related files to chatroom package (#42)
