+++ app/src/main/java/com/example/rentit/feature/MainView.kt
@@ -34,9 +34,9 @@ import com.example.rentit.common.component.moveScreen
 import com.example.rentit.common.theme.Gray400
 import com.example.rentit.common.theme.PrimaryBlue500
 import com.example.rentit.common.theme.RentItTheme
-import com.example.rentit.feature.chat.AcceptConfirmationScreen
+import com.example.rentit.feature.chat.chatroom.requestaccept.confirm.RequestAcceptConfirmScreen
 import com.example.rentit.feature.chat.ChatListScreen
-import com.example.rentit.feature.chat.ChatroomScreen
+import com.example.rentit.feature.chat.chatroom.ChatroomScreen
 import com.example.rentit.feature.home.HomeScreen
 import com.example.rentit.feature.mypage.MyPageScreen
 import com.example.rentit.feature.productdetail.reservation.request.ResvRequestScreen
@@ -185,7 +185,7 @@ fun ChatroomNavHost(productId: Int?, reservationId: Int?, chatRoomId: String?) {
     val navHostController: NavHostController = rememberNavController()
     NavHost(navController =  navHostController, startDestination = NavigationRoutes.CHATROOM){
         composable(NavigationRoutes.CHATROOM) { ChatroomScreen(navHostController, productId, reservationId, chatRoomId) }
-        composable(NavigationRoutes.ACCEPTCONFIRM) { AcceptConfirmationScreen(navHostController) }
+        composable(NavigationRoutes.ACCEPTCONFIRM) { RequestAcceptConfirmScreen(navHostController) }
         composable(NavigationRoutes.MAIN) { MainView() }
     }
 }
